added userService delete user by role and own id

+++ src/modules/user/user.service.ts
@@ -1,6 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { CreateUserDto, UpdateUserDto } from '@modules/user/dto';
 import { UserRepository } from './repository/user.repository';
+import { IJwtPayload } from '@modules/auth/interfaces';
+import { E_UserRole, $Enums } from '@prisma/client';
 
 @Injectable()
 export class UserService {
@@ -20,7 +22,15 @@ export class UserService {
   update(id: string, updateUserDto: UpdateUserDto) {
     return this.userRepository.updateUser(id, updateUserDto);
   }
-  remove(id: string) {
+  remove(id: string, user: IJwtPayload) {
+    console.log('🚀 ~ UserService ~ remove ~ user:', user);
+    //COMMENT delete only own user and with status user role ADMIN
+    const isAdmin = user.roles
+      ? user.roles.includes($Enums.E_UserRole.ADMIN)
+      : false;
+    if (id !== user.id && !isAdmin) {
+      throw new ForbiddenException();
+    }
     return this.userRepository.deleteUser(id);
   }
 }
